Move typelevel-related things to new extra module
